feat(graphic settings): move shadow map resolution to graphic settings (#8267)
