+++ packages/editor/src/components/properties/LightShadowProperties.tsx
@@ -32,36 +32,9 @@ import { Component, useComponent } from '@etherealengine/engine/src/ecs/function
 import BooleanInput from '../inputs/BooleanInput'
 import InputGroup from '../inputs/InputGroup'
 import NumericInputGroup from '../inputs/NumericInputGroup'
-import SelectInput from '../inputs/SelectInput'
-import { updateProperties, updateProperty } from './Util'
+import { updateProperty } from './Util'
 
-/**
- *  Array containing options for shadow resolution
- */
-const ShadowMapResolutionOptions = [
-  {
-    label: '256px',
-    value: 256
-  },
-  {
-    label: '512px',
-    value: 512
-  },
-  {
-    label: '1024px',
-    value: 1024
-  },
-  {
-    label: '2048px',
-    value: 2048
-  },
-  {
-    label: '4096px (not recommended)',
-    value: 4096
-  }
-]
-
-//creating properties for LightShadowProperties component
+/**creating properties for LightShadowProperties component */
 type LightShadowPropertiesProps = {
   entity: Entity
   comp: Component<any, any>
@@ -70,15 +43,9 @@ type LightShadowPropertiesProps = {
 /**
  * OnChangeShadowMapResolution used to customize properties of LightShadowProperties
  * Used with LightNodeEditors.
- *
- * @type {[class component]}
  */
 export const LightShadowProperties = (props: LightShadowPropertiesProps) => {
   const { t } = useTranslation()
-
-  const changeShadowMapResolution = (resolution) => {
-    updateProperties(props.comp, { shadowMapResolution: resolution })
-  }
 
   const lightComponent = useComponent(props.entity, props.comp).value as any
 
@@ -86,14 +53,6 @@ export const LightShadowProperties = (props: LightShadowPropertiesProps) => {
     <>
       <InputGroup name="Cast Shadows" label={t('editor:properties.directionalLight.lbl-castShadows')}>
         <BooleanInput value={lightComponent.castShadow} onChange={updateProperty(props.comp, 'castShadow')} />
-      </InputGroup>
-      <InputGroup name="Shadow Map Resolution" label={t('editor:properties.directionalLight.lbl-shadowmapResolution')}>
-        <SelectInput
-          key={props.entity}
-          options={ShadowMapResolutionOptions}
-          value={lightComponent.shadowMapResolution?.x}
-          onChange={changeShadowMapResolution}
-        />
       </InputGroup>
       <NumericInputGroup
         name="Shadow Bias"
